backend/requirements.txt
# LeadScraper AI - Python Dependencies
# Python 3.11+ required

# ============================================
# Core Framework
# ============================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# ============================================
# Database (Supabase)
# ============================================
supabase==2.9.1  # Updated to avoid httpx conflicts
psycopg2-binary==2.9.9  # PostgreSQL adapter (optional, for direct DB access)

# ============================================
# HTTP Clients
# ============================================
httpx>=0.24.0  # Async HTTP client (better for FastAPI than requests)
requests==2.31.0  # Synchronous HTTP client (for some APIs)

# ============================================
# Environment & Configuration
# ============================================
python-dotenv==1.0.0

# ============================================
# Google APIs
# ============================================
google-generativeai==0.3.2  # Gemini AI
google-api-python-client==2.110.0  # PageSpeed Insights API

# ============================================
# Web Scraping
# ============================================
beautifulsoup4==4.12.3
lxml==5.1.0  # HTML parser for BeautifulSoup
playwright==1.41.0  # Optional: For JavaScript-heavy sites
aiohttp==3.9.1  # Async HTTP for web scraping

# ============================================
# Utilities
# ============================================
python-multipart==0.0.6  # For form data parsing
python-jose[cryptography]==3.3.0  # JWT tokens (if needed for auth)
email-validator==2.1.0  # Email validation

# ============================================
# Development & Testing (Optional)
# ============================================
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0  # For testing FastAPI endpoints
black==24.2.0  # Code formatter
ruff==0.2.1  # Linter
